Refactored Print window
