+++ source/print/index.htm
@@ -21,48 +21,11 @@
     </style>
 </head>
 <body>
-  <div class="print-window">
-    <div class="title">
-      <h1></h1>
-      <!-- TODO: Not yet working -->
-      <div id="init-print"><clr-icon shape="printer"></clr-icon></div>
-    </div>
-    <div class="content">
-    </div>
-  </div>
+  <div id="toolbar"></div>
+
+  <div id="app"></div>
 
   <!-- This is the menu bar -- only visible on Windows & Linux -->
   <div id="menubar"></div>
-  <!--
-      Windows window controls, taken from
-      https://github.com/sidneys/electron-titlebar-windows (License: MIT)
-    -->
-
-  <div class="window-controls">
-    <div class="minimise">
-      <svg x="0px" y="0px" viewBox="0 0 10 1">
-        <rect fill="#000000" width="10" height="1"></rect>
-      </svg>
-    </div>
-    <div class="resize">
-      <svg class="fullscreen-svg" x="0px" y="0px" viewBox="0 0 10 10">
-        <path fill="#000000" d="M 0 0 L 0 10 L 10 10 L 10 0 L 0 0 z M 1 1 L 9 1 L 9 9 L 1 9 L 1 1 z "/>
-      </svg>
-
-      <svg class="maximise-svg" x="0px" y="0px" viewBox="0 0 10 10">
-        <mask id="Mask">
-          <rect fill="#FFFFFF" width="10" height="10"></rect>
-          <path fill="#000000" d="M 3 1 L 9 1 L 9 7 L 8 7 L 8 2 L 3 2 L 3 1 z"/>
-          <path fill="#000000" d="M 1 3 L 7 3 L 7 9 L 1 9 L 1 3 z"/>
-        </mask>
-        <path fill="#000000" d="M 2 0 L 10 0 L 10 8 L 8 8 L 8 10 L 0 10 L 0 2 L 2 2 L 2 0 z" mask="url(#Mask)"/>
-      </svg>
-    </div>
-    <div class="close">
-      <svg x="0px" y="0px" viewBox="0 0 10 10">
-        <polygon fill="#000000" points="10,1 9,0 5,4 1,0 0,1 4,5 0,9 1,10 5,6 9,10 10,9 6,5"></polygon>
-      </svg>
-    </div>
-  </div>
 </body>
 </html>
